Use explict int sizes

+++ src/eg3dmath.c
@@ -19,39 +19,39 @@
 /* sin/cos via table lookup + linear interpolation.
  * Mirrors sineLookup: idx = angle>>8, frac = angle&0xFF, then
  *   result = lut[idx] + round((lut[idx+1] - lut[idx]) * frac / 256). */
-static int sineInterp(unsigned angle) {
-    int idx = (angle >> 8) & 0xFF;
-    int frac = angle & 0xFF;
-    int v0 = g_angleLut[idx];
-    int v1 = g_angleLut[idx + 1];
-    long step = (long)(v1 - v0) * frac;
-    return v0 + (int)((step + 0x80) >> 8);
+static int16 sineInterp(uint16 angle) {
+    int16 idx = (angle >> 8) & 0xFF;
+    int16 frac = angle & 0xFF;
+    int16 v0 = g_angleLut[idx];
+    int16 v1 = g_angleLut[idx + 1];
+    int32 step = (int32)(v1 - v0) * frac;
+    return v0 + (int16)((step + 0x80) >> 8);
 }
 
-int sine(int angle) {
+int16 sine(int16 angle) {
     return sineInterp(angle);
 }
 
-int cosine(int angle) {
+int16 cosine(int16 angle) {
     /* cos(x) = sin(x + 90deg); a quarter turn is 0x4000 in 16-bit angle space. */
-    return sineInterp((unsigned)angle + 0x4000);
+    return sineInterp((uint16)angle + 0x4000);
 }
 
 /* Q15-style fixed multiply: returns round((a*b) >> 15).
  * Replicates the exact shl/rcl/adc sequence of the ASM fixedMulQ14:
  *   P = a*b; result = (P>>15) + (bit14 of P). */
-int fixedMulQ14(int a, int b) {
-    long p = (long)a * (long)b;
-    return (int)((p >> 15) + ((p >> 14) & 1L));
+int16 fixedMulQ14(int16 a, int16 b) {
+    int32 p = (int32)a * (int32)b;
+    return (int16)((p >> 15) + ((p >> 14) & 1L));
 }
 
 /* In-place 32-bit shifts (pascal: args pushed left-to-right, callee cleans up).
- * The ASM dispatches to the MSC long-shift helpers; >>= on a signed long is the
+ * The ASM dispatches to the MSC long-shift helpers; >>= on a signed int32 is the
  * arithmetic shift those helpers perform. */
-void pascal shiftLongLeftInPlace(int count, long *ptr) {
+void pascal shiftLongLeftInPlace(int16 count, int32 *ptr) {
     *ptr <<= count;
 }
 
-void pascal shiftLongRightInPlace(int count, long *ptr) {
+void pascal shiftLongRightInPlace(int16 count, int32 *ptr) {
     *ptr >>= count;
 }
